# docs/swagger.yaml
basePath: /
definitions:
  github_com_mbr_moviesx_internal_store.Actor:
    properties:
      birthYear:
        type: integer
      deathYear:
        type: integer
      id:
        type: string
      movies:
        items:
          $ref: '#/definitions/github_com_mbr_moviesx_internal_store.ActorMovie'
        type: array
      name:
        type: string
      profession:
        items:
          type: string
        type: array
    type: object
  github_com_mbr_moviesx_internal_store.ActorMovie:
    properties:
      movieId:
        type: string
      title:
        type: string
    type: object
  github_com_mbr_moviesx_internal_store.MovieDetail:
    properties:
      genres:
        items:
          type: string
        type: array
      id:
        type: string
      rating:
        type: number
      roles:
        items:
          $ref: '#/definitions/github_com_mbr_moviesx_internal_store.Role'
        type: array
      runtime:
        type: integer
      title:
        type: string
      votes:
        type: integer
      year:
        type: integer
    type: object
  github_com_mbr_moviesx_internal_store.MovieListItem:
    properties:
      cast:
        items:
          $ref: '#/definitions/github_com_mbr_moviesx_internal_store.Role'
        type: array
      genres:
        items:
          type: string
        type: array
      id:
        type: string
      rating:
        type: number
      runtime:
        type: integer
      title:
        type: string
      votes:
        type: integer
      year:
        type: integer
    type: object
  github_com_mbr_moviesx_internal_store.Page-github_com_mbr_moviesx_internal_store_Actor:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_mbr_moviesx_internal_store.Actor'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  github_com_mbr_moviesx_internal_store.Page-github_com_mbr_moviesx_internal_store_MovieListItem:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_mbr_moviesx_internal_store.MovieListItem'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  github_com_mbr_moviesx_internal_store.Role:
    properties:
      actorId:
        type: string
      category:
        type: string
      characters:
        items:
          type: string
        type: array
      job:
        type: string
      name:
        type: string
      order:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: Read-only movie catalog API
  title: moviesx API
  version: 0.4.0
paths:
  /api/actors:
    get:
      parameters:
      - description: Name search substring (min 2 chars)
        in: query
        name: q
        type: string
      - description: Page number (default 1)
        in: query
        name: pageNumber
        type: integer
      - description: Page size (default 20, max 100)
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mbr_moviesx_internal_store.Page-github_com_mbr_moviesx_internal_store_Actor'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List actors
      tags:
      - actors
  /api/actors/{id}:
    get:
      parameters:
      - description: Actor ID (e.g. nm0000001)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mbr_moviesx_internal_store.Actor'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get actor by ID
      tags:
      - actors
  /api/genres:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      summary: List all genres
      tags:
      - genres
  /api/movies:
    get:
      parameters:
      - description: Title search substring (min 2 chars)
        in: query
        name: q
        type: string
      - description: Genre filter
        in: query
        name: genre
        type: string
      - description: Release year
        in: query
        name: year
        type: integer
      - description: Minimum rating
        in: query
        name: rating
        type: number
      - description: Filter by actor ID
        in: query
        name: actorId
        type: string
      - description: Page number (default 1)
        in: query
        name: pageNumber
        type: integer
      - description: Page size (default 20, max 100)
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mbr_moviesx_internal_store.Page-github_com_mbr_moviesx_internal_store_MovieListItem'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List movies
      tags:
      - movies
  /api/movies/{id}:
    get:
      parameters:
      - description: Movie ID (e.g. tt0000001)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mbr_moviesx_internal_store.MovieDetail'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get movie by ID
      tags:
      - movies
  /healthz:
    get:
      produces:
      - text/plain
      responses:
        "200":
          description: pass
          schema:
            type: string
      summary: Liveness check
      tags:
      - system
  /readyz:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Readiness check
      tags:
      - system
  /version:
    get:
      produces:
      - text/plain
      responses:
        "200":
          description: version string
          schema:
            type: string
      summary: Get server version
      tags:
      - system
schemes:
- http
swagger: "2.0"